libcopyfile: Fix `copyfile_open`
On macOS, depending on whether or not the `COPYFILE_NOFOLLOW_SRC`
flag is set, either `lstatx_np` or `statx_np` is called.

aquaBSD however doesn't have such functions in its standard library,
so I'll have to eventually come back to this and rewrite
`copyfile_open` "properly" (or even better, rewrite the entirety of
libcopyfile and perhaps merge upstream with FreeBSD).